refactor: move JoinHouseholdRequest to dto.request package and update imports
