+++ src/test/java/com/group7/krisefikser/repository/JoinHouseholdRequestRepoTest.java
@@ -1,6 +1,6 @@
 package com.group7.krisefikser.repository;
 
-import com.group7.krisefikser.model.JoinHouseholdRequest;
+import com.group7.krisefikser.dto.request.JoinHouseholdRequest;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
